requirements.txt
streamlit
pickle4
langchain
load-dotenv
faiss-cpu
google-generativeai
langchain_google_genai
unstructured
libmagic
cffi
python_magic_bin
transformers
langchain-community